==> CdSe/1. Scaling and encoding.ipynb <==
import numpy as np
import pandas as pd
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder

#Turns the CSV file into a pandas dataframe
df = pd.read_csv('dataset_CdSe_raw.csv')
df

#Identifies the shape of the original dataframe
df.shape

#Checks if there are any columns with no values
df.isna().sum()

#Separate out initial DataFrame into the input features and output features
df_input = df.drop(columns =['Chalcogen_source','diameter_nm', 'abs_nm', 'emission_nm', 'Diameter from', 'Citation'], inplace = False, axis = 1) #Dropped out FWHM, Phase and PLQY columns as well because they have NaN values
df_output = df[['diameter_nm', 'abs_nm', 'emission_nm']]

#Checks the column names, and ensures that they do not have any leading or trailing spaces
df_input.columns = df_input.columns.str.strip()
df_output.columns = df_output.columns.str.strip()

df_input.info()

#Converts the values in the "Temperature" and "Growth" Columns into float types
df_input['Injection Temp (Celsius)'] = df_input['Injection Temp (Celsius)'].astype(float)
df_input['Growth Temp (Celsius)'] = df_input['Growth Temp (Celsius)'].astype(float)

df_input.info() 

#Initializes 2 lists to contain all of the numerical and categorical input columns
input_num_cols = [col for col in df_input.columns if df[col].dtypes !='O']
input_cat_cols = [col for col in df_input.columns if df[col].dtypes =='O']

input_cat_cols

#Initializes the ColumnTransformer object, and specifies what it will do with a passed in dataframe
ct = ColumnTransformer([
    ('step1', StandardScaler(), input_num_cols),
    ('step2', OneHotEncoder(sparse=False, handle_unknown='ignore'), input_cat_cols)
], remainder = 'passthrough')

#Uses the ColumnTransformer object to modify the input columns
df_input_scaled_encoded = pd.DataFrame(ct.fit_transform(df_input))
df_input_scaled_encoded

#### Now that we have transformed all of the relevant columns, how can we take back all of the relevant column names?

ct.transformers_

#Look into ct object to see the column titles
array_of_cat_titles = ct.transformers_[1][1].get_feature_names()
len(array_of_cat_titles) #There are a total of 41 encoded categorical columns

#Number of numerical columns in the dataset
len(input_num_cols)

Pausing to do an initial sanity check here, we noticed that our initial dataset had 20 numerical columns (we checked the length of our `input_num_cols` array). 

Our from two cells above, we notice that our one hot encoding of the categorical columns resulted in 41 columns.

If we add 41 + 20, we obtain 61, which is consistent with the number of columns displayed when we printed `df_input_scaled_encoded`. 

We note here that the one hot encoded categorical columns are all appended at the end, and this is apparent when we displayed `df_input_scaled_encoded`. 

What we need to do next is figure out which column index corresponds to which category. E.g Column 30 refers to which category?


for i in range(len(input_num_cols)):
    df_input_scaled_encoded.rename(columns={df_input_scaled_encoded.columns[i]: input_num_cols[i]}, inplace = True)

for j in range(len(array_of_cat_titles)):
    df_input_scaled_encoded.rename(columns={df_input_scaled_encoded.columns[i+1]: array_of_cat_titles[j]}, inplace = True)
    i = i + 1


df_input_scaled_encoded.columns[43]

df_input_scaled_encoded

df_scaled_encoded = pd.concat([df_input_scaled_encoded, df_output], axis = 1)
df_scaled_encoded

df_scaled_encoded.to_csv('dataset_CdSe_scaled.csv')

df_scaled_encoded

#Saves the row indexes to drop for absorbance modelling into a list
total_row_num = len(df_scaled_encoded)
drop_list_abs =[]
for row_i in range(total_row_num):
    if df_scaled_encoded['abs_nm'].values[row_i] == 'None':
        drop_list_abs.append(row_i)
    
len(drop_list_abs)


#Drops the appropriate rows
df_absorbance_scaled_encoded = df_scaled_encoded.drop(drop_list_abs)

#Saves the data for absorbance modelling to CSV
df_absorbance_scaled_encoded.to_csv('dataset_CdSe_abs_filler.csv')

Drop unuseful columns **MANUALLY**


#Saves the row indexes to drop for PL modelling into a list
total_row_num = len(df_scaled_encoded)
drop_list_pl =[]
for row_i in range(total_row_num):
    if df_scaled_encoded['emission_nm'].values[row_i] == 'None':
        drop_list_pl.append(row_i)
    
len(drop_list_pl)

#Drops the appropriate rows
df_pl_scaled_encoded = df_scaled_encoded.drop(drop_list_pl)

#Saves the data for PL modelling to CSV
df_pl_scaled_encoded.to_csv('dataset_CdSe_emission_filler.csv')

Drop unuseful columns **MANUALLY**
